Select decoded id on focus of the encode input field
I will be using this to predominantly decode encoded ID strings in the
format NodeType:1 etc. I want this to focus on the decoded int by
default so I can copy it. The decode input field remains selecting
the whole encoded ID as that is what I'd want to copy.